+++ packages/dapp/components/redeem/debt-coupon.tsx
@@ -2,9 +2,10 @@ import { constrainNumber, formatTimeDiff } from "@/lib/utils";
 import withLoadedContext, { LoadedContext } from "@/lib/with-loaded-context";
 import { BigNumber, ethers } from "ethers";
 import { ChangeEvent, Dispatch, memo, SetStateAction, useEffect, useMemo, useState } from "react";
-import useBalances, { Balances } from "../lib/hooks/use-balances";
+import useBalances from "../lib/hooks/use-balances";
 import useTransactionLogger from "../lib/hooks/use-transaction-logger";
 import usePrices from "./lib/use-prices";
+import { Balances } from "../lib/types";
 
 type Actions = {
   onRedeem: () => void;
